Convert hashtags into component
