+++ apps/builddao/widget/events/ListView.jsx
@@ -1,3 +1,7 @@
+const { Hashtag } = VM.require("buildhub.near/widget/components") || {
+  Hashtag: () => <></>,
+};
+
 const events = props.events ?? [];
 const currentDate = props.currentDate;
 
@@ -72,38 +76,6 @@ const StyledEvent = styled.div`
     }
   }
 `;
-
-const StyledHashtag = styled.span`
-  display: flex;
-  padding: 4px 8px;
-  justify-content: center;
-  align-items: center;
-  align-content: center;
-  gap: 4px;
-  flex-wrap: wrap;
-
-  border-radius: 2px;
-  border: 1px solid var(--Yellow, #ffaf51);
-
-  color: var(--White-100, #fff);
-
-  /* Body/10px */
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-
-  .tag {
-    color: var(--Yellow, #ffaf51);
-  }
-`;
-const Hashtag = ({ children }) => {
-  return (
-    <StyledHashtag>
-      <span className="tag">#</span> {children}
-    </StyledHashtag>
-  );
-};
 
 const formatStartTime = (time) => {
   const date = new Date(time);
